refactor(tui::ui::components::popups::error): move function "mount_error_popup" to popup file
