+++ tui/src/ui/components/mod.rs
@@ -53,7 +53,7 @@ pub use music_library::MusicLibrary;
 pub use playlist::Playlist;
 pub use podcast::{EpisodeList, FeedsList};
 pub use popups::{
-    ErrorPopup, HelpPopup, MessagePopup, PodcastAddPopup, QuitPopup, SavePlaylistConfirmPopup,
+    HelpPopup, MessagePopup, PodcastAddPopup, QuitPopup, SavePlaylistConfirmPopup,
     SavePlaylistPopup,
 };
 pub use progress::Progress;
